user-study labels

+++ data_generator.py
@@ -2,7 +2,7 @@ import json
 import os
 import numpy as np
 import cv2
-import tqdm
+from tqdm import tqdm
 import copy
 import random
 import torch
@@ -61,8 +61,9 @@ def create_frame_session_folders(data_config):
             participant_path = os.path.join(data_root_dir, participant_type, participant)
             config_path = os.path.join(participant_path, "config.json")
             print(participant)
-            os.mkdir(os.path.join(labels_root, participant))
-            os.mkdir(os.path.join(frames_root, participant))
+            if not os.path.exists(os.path.join(labels_root, participant)):
+                os.mkdir(os.path.join(labels_root, participant))
+                os.mkdir(os.path.join(frames_root, participant))
 
             participant_labels_folder = os.path.join(labels_root, participant)
             participant_frames_folder = os.path.join(frames_root, participant)
@@ -74,34 +75,37 @@ def create_frame_session_folders(data_config):
                     syncing_poses = ground_truths.get("syncing_poses", [])
                     print(videos, syncing_poses)
                     for i, syncing_pose in enumerate(syncing_poses):
-                        print("Processing Session " + str(i))
-                        video = videos[i]
-                        
-                        os.mkdir(os.path.join(participant_labels_folder, f"session_{i+1:02d}"))
-                        labels_folder = os.path.join(participant_labels_folder, f"session_{i+1:02d}")
-                        
-                        os.mkdir(os.path.join(participant_frames_folder, f"session_{i+1:02d}"))
-                        frames_folder = os.path.join(participant_frames_folder, f"session_{i+1:02d}")
-    
-                        video_path = os.path.join(participant_path, video) 
-                        cap = cv2.VideoCapture(video_path)
-                        frame_count = 1
-                        while True:
-                            ret, frame = cap.read()
-                            if not ret:
-                                break
-                            frame_path = os.path.join(frames_folder, f"participant_{participant}_session_{i+1:02d}_frame_{frame_count:04d}.jpg")
-                            cv2.imwrite(frame_path, frame)
-                            frame_count += 1
-                        cap.release()
-                        
-                        labels_path = os.path.join(labels_folder, "labels.txt")
-                        with open(labels_path, 'w') as labels_file:
-                            frame_files = os.listdir(frames_folder)
-                            for frame_index, frame_file in enumerate(frame_files):
-                                frame_file_path = os.path.abspath(os.path.join(frames_folder, frame_file))
-                                label = 1 if frame_index+1 == syncing_pose else 0
-                                labels_file.write(f"{label} {frame_file_path}\n")
+                        if not os.path.exists(os.path.join(participant_labels_folder, f"session_{i+1:02d}")):
+                            print("Processing Participant " + participant + " Session " + str(i+1))
+                            video = videos[i]
+                            
+                            os.mkdir(os.path.join(participant_labels_folder, f"session_{i+1:02d}"))
+                            labels_folder = os.path.join(participant_labels_folder, f"session_{i+1:02d}")
+                            
+                            os.mkdir(os.path.join(participant_frames_folder, f"session_{i+1:02d}"))
+                            frames_folder = os.path.join(participant_frames_folder, f"session_{i+1:02d}")
+        
+                            video_path = os.path.join(participant_path, video) 
+                            cap = cv2.VideoCapture(video_path)
+                            frame_count = 1
+                            while True:
+                                ret, frame = cap.read()
+                                if not ret:
+                                    print("No frame at frame: " + str(frame_count))
+                                    break
+                                frame = cv2.resize(frame, (224, 224))
+                                frame_path = os.path.join(frames_folder, f"participant_{participant}_session_{i+1:02d}_frame_{frame_count:04d}.jpg")
+                                cv2.imwrite(frame_path, frame)
+                                frame_count += 1
+                            cap.release()
+                            
+                            labels_path = os.path.join(labels_folder, "labels.txt")
+                            with open(labels_path, 'w') as labels_file:
+                                frame_files = os.listdir(frames_folder)
+                                for frame_index, frame_file in enumerate(frame_files):
+                                    frame_file_path = os.path.abspath(os.path.join(frames_folder, frame_file))
+                                    label = 1 if frame_index+1 == syncing_pose else 0
+                                    labels_file.write(f"{label} {frame_file_path}\n")
 
 
 if __name__ == "__main__":
